backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic-settings
python-dotenv
alembic
sqlalchemy
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
pyotp
pandas
openpyxl
python-multipart
workalendar
qrcode[pil]
cryptography
pypdf
lxml